Tauri runtime error handling (#18)
* Fixed Tauri runtime error handling
* Removed unnecessary runtime code
* Fixed Windows compilation and instructions

---------
