+++ Cargo.toml
@@ -26,6 +26,7 @@ name = "rezasm_wasm"
 
 
 [build-dependencies]
+cargo-watch = "8.4.1"
 tauri-build = { version = "1.4.0", features = [] }
 wasm-pack = "0.12.1"
 
@@ -46,8 +47,8 @@ lazy_static = "1.4.0"
 regex = "1.8.4"
 serde = { version = "1.0.188", features = ["derive"] }
 serde_json = "1.0.105"
+tauri-cli = "1.4.0"
 thiserror = "1.0.47"
-tokio = { version = "1.32.0", features = ["sync", "rt"] }
 wasm-bindgen = "0.2.87"
 
 [workspace]
