+++ README.md
@@ -11,28 +11,31 @@ The instructions would be intuitive and simple compared to MIPS (e.g., no system
 ## Getting Started (Windows)
 1. Ensure you have installed [Rust](https://www.rust-lang.org/learn/get-started) and [Node.js](https://nodejs.org/en/download)
 2. Download the [Visual Studio Installer](https://visualstudio.microsoft.com/downloads/) and install the Desktop Development with C++ Workload
-3. Install cargo-tauri using `cargo install tauri-cli` in a terminal emulator
-4. Install tailwindcss using `npm install -g tailwindcss` in a terminal emulator
-5. Clone this repository to your system
-6. Open this repository in your terminal emulator
-7. Run `npm i` in a terminal emulator to install the node dependencies
-8. Run the application in development mode using `cargo tauri dev`
+3. Install cargo-tauri using `cargo install tauri-cli wasm-pack` in a terminal emulator
+4. Clone this repository to your system
+5. Open this repository in your terminal emulator
+6. Update npm using `npm install -g npm@latest`
+7. Run `npm install` in a terminal emulator to install the node dependencies
+8. Run the tauri application in development mode using `cargo tauri dev`
+9. Test the WebAssembly code by going to http://localhost:1420 while the tauri application is open
 
 ## Getting Started (Linux)
 1. Ensure you have installed [Rust](https://www.rust-lang.org/learn/get-started) and [Node.js](https://nodejs.org/en/download)
-2. Install cargo-tauri using `cargo install tauri-cli`
-3. Install tailwindcss using `npm install -g tailwindcss`
-4. Clone this repository to your system
-5. Open this repository in your terminal
-6. Run `npm i` in your terminal to install the node dependencies
-7. Run the application in development mode using `cargo tauri dev`
+2. Install cargo-tauri using `cargo install tauri-cli wasm-pack`
+3. Clone this repository to your system
+4. Open this repository in your terminal
+5. Update npm using `npm install -g npm@latest`
+6. Run `npm install` in your terminal to install the node dependencies
+7. Run the tauri application in development mode using `cargo tauri dev`
+8. Test the WebAssembly code by going to http://localhost:1420 while the tauri application is open
 
 ## Getting Started (MacOS)
 **Under Construction / May not work yet**
 1. Ensure you have installed [Rust](https://www.rust-lang.org/learn/get-started) and [Node.js](https://nodejs.org/en/download)
-2. Install cargo-tauri using `cargo install tauri-cli`
-3. Install tailwindcss using `npm install -g tailwindcss`
-4. Clone this repository to your system
-5. Open this repository in your terminal
-6. Run `npm i` in your terminal to install the node dependencies
-7. Run the application in development mode using `cargo tauri dev`
+2. Install cargo-tauri using `cargo install tauri-cli wasm-pack`
+3. Clone this repository to your system
+4. Open this repository in your terminal
+5. Update npm using `npm install -g npm@latest`
+6. Run `npm install` in your terminal to install the node dependencies
+7. Run the tauri application in development mode using `cargo tauri dev`
+8. Test the WebAssembly code by going to http://localhost:1420 while the tauri application is open
