+++ package.json
@@ -4,16 +4,16 @@
   "version": "0.1.0",
   "type": "module",
   "scripts": {
-    "dev": "npm-run-all -p vite-dev tailwind-dev",
-    "build": "npm-run-all -p vite-build tailwind-build",
+    "dev": "npm-run-all -p tailwind-dev vite-dev",
+    "build": "npm run tailwind-build && npm run vite-build ",
     "preview": "vite preview",
     "format": "eslint --fix",
-    "tauri-dev": "tauri dev",
-    "tauri-build": "tauri build",
     "tailwind-dev": "tailwindcss -i ./src/styles.css -o ./dist/output.css --watch",
     "tailwind-build": "tailwindcss -i ./src/styles.css -o ./dist/output.css",
-    "vite-dev": "vite dev",
-    "vite-build": "vite build"
+    "vite-dev": "npm run wasm-dev && vite dev",
+    "vite-build": "npm run wasm-build && vite build",
+    "wasm-dev": "wasm-pack build -d ../../dist/wasm --target web --dev rezasm-app/rezasm-wasm/",
+    "wasm-build": "wasm-pack build -d ../../dist/wasm --target web --dev rezasm-app/rezasm-wasm/"
   },
   "dependencies": {
     "@tauri-apps/api": "^1.4.0",
@@ -21,6 +21,7 @@
     "react-dom": "^18.2.0",
     "vite-plugin-top-level-await": "^1.3.1",
     "vite-plugin-wasm": "^3.2.2",
+    "vite-plugin-wasm-pack": "^0.1.12",
     "webpack-cli": "^5.1.4",
     "webworker-promise": "^0.5.1"
   },
