+++ rezasm-app/rezasm-tauri/Cargo.toml
@@ -21,7 +21,6 @@ lazy_static = "1.4.0"
 serde = { version = "1.0.188", features = ["derive"] }
 serde_json = "1.0.105"
 tauri = { version = "1.4.1", features = ["shell-open"] }
-tokio = { version = "1.32.0", features = ["sync", "rt"] }
 
 # this feature is used for production builds or when `devPath` points to the filesystem
 # DO NOT REMOVE!!
