+++ rezasm-app/rezasm-tauri/icons/icon.ico
