+++ rezasm-app/rezasm-tauri/tauri.conf.json
@@ -8,7 +8,7 @@
   },
   "package": {
     "productName": "rezasm-web",
-    "version": "0.1.0"
+    "version": "1.0.0"
   },
   "tauri": {
     "allowlist": {
@@ -24,7 +24,11 @@
     "bundle": {
       "active": true,
       "targets": "all",
-      "identifier": "org.ezasm.rust"
+      "identifier": "org.ezasm.rust",
+      "icon": [
+        "icons/icon.png",
+        "icons/icon.ico"
+      ]
     },
     "security": {
       "csp": null
