+++ rezasm-app/rezasm-wasm/Cargo.toml
@@ -19,5 +19,4 @@ rezasm-macro = { path = "../../rezasm-source/rezasm-macro" }
 rezasm-web-core = { path = "../../rezasm-source/rezasm-web-core" }
 
 lazy_static = "1.4.0"
-tokio = { version = "1.32.0", features = ["sync", "rt"] }
 wasm-bindgen = "0.2.87"
