--- rezasm-source/rezasm-web-core/src/util/mod.rs
@@ -1,2 +1,0 @@
-pub mod commands;
-pub mod runtime;
